Refactored tests to extract testcase.
